requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
faiss-cpu>=1.7.4
openai>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
